// src/app/models/user.ts
export class User {
  activated?: boolean;
  activated_at?: string;
  activation_digest?: string;
  admin?: boolean;
  created_at?: string;
  email?: string;
  id?: number;
  name?: string;
  password_digest?: string;
  remember_digest?: string;
  reset_digest?: string;
  reset_sent_at?: string;
  updated_at?: string;
}